Fix build config
.csproj file has been modified to restore some of old build
functionality, as this needs to be in .csproj instead of .csproj.user.
Build script refactored to always copy files (will have no effect for
debug builds if output directory is the project directory due to /D),
but now works better for output directories overridden in .csproj.user.

Works on L0laapk3/RimWorld_FinishTheDamnThing#2.